test: move to new destination
